program prints first 50 fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,24 +7,24 @@
 int main(void)
 {
 	int n1 = 0;
-	int n2 = 1;
-	int n3 = 2;
+	long n2 = 1;
+	long n3 = 2;
 
 	while (n1 < 50)
 	{
 		if (n1 == 0)
 		{
-			printf("%d", n2);
+			printf("%ld", n2);
 		}
 		else if (n1 == 1)
 		{
-			printf(", %d", n3);
+			printf(", %ld", n3);
 		}
 		else
 		{
 			n3 = n3 + n2;
 			n2 = n3 - n2;
-			printf(", %d", n3);
+			printf(", %ld", n3);
 		}
 		n1++;
 	}
